requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==2.1
django-paystack==0.3
future==0.17.1
idna==2.7
Pillow==5.4.0
pytz==2017.2
requests==2.20.0
urllib3==1.24.1
